Large refactor of convolution / Convolver to get the code running fast with numba, including:
- Frame Arrays are now a 2d numpy array instead of list of numpy arrays.
- The Kernel value of each frame array value is stored in a separate
array, and is used to eliminate scanning over -1s.
- Separate Coonvolvers for images / mapping matrices are now produced
by the masked image.

+++ src/imaging/scaled_array.py
@@ -14,7 +14,7 @@ class ScaledArray(np.ndarray):
     """
 
     def __new__(cls, array, pixel_scale=1, *args, **kwargs):
-        return np.array(array).view(cls)
+        return np.array(array, dtype='float64').view(cls)
 
     # noinspection PyUnusedLocal
     def __init__(self, array, pixel_scale=1):
